Improved graph redrawing.
-Graps are now only redrawn when switching from table to graph view if the
user was in table view as the graph was being loaded. Otherwise, they are
simply resized to take care of the cursor issue.

-Things like vm.sdWithinTab.config weren't getting reset on every request.
Now they are.

-Fixed the graph summary not updating to show new info. This was due to
one-time binding.

-Still have to fix memory leak.

+++ app/js/controllers/mainGraphTab/MGQueryController.js
@@ -105,6 +105,7 @@
          * @memberOf controllers.MGQueryController
          */
         function refreshGraph(filter) {
+            GraphConfigService.destroyGraph(vm);
             vm.clearLocatedGene();
             MGSharedData.resetWTM(vm);
             getConfigForGraph(filter);
@@ -189,12 +190,18 @@
             }
             return null;
         }, function(newValue, oldValue) {
-            if (newValue == vm.displayModes.graph && newValue != oldValue) {
+            if (newValue == vm.displayModes.graph && newValue != oldValue && vm.needsRedraw) {
                 $timeout(function() {
                     if (vm.sdWithinTab.config != null && vm.sdWithinTab.cy != null) {
                         GraphConfigService.destroyGraph(vm);
                         vm.needsRedraw = false;
                         vm.sdWithinTab.cy = GraphConfigService.applyConfig(vm, vm.sdWithinTab.config, "cyMain");
+                    }
+                }, 250);
+            } else if (newValue == vm.displayModes.graph && newValue != oldValue && !vm.needsRedraw) {
+                $timeout(function() {
+                    if (vm.sdWithinTab.config != null && vm.sdWithinTab.cy != null) {
+                        vm.resize(vm);
                     }
                 }, 250);
             }
